Sesiones/S10/test_pila.py
# -*- coding: utf-8 -*-
"""
Created on Wed Dec 13 10:03:24 2017

@author: Mario
"""

import unittest
from pila import *

class TestMipila(unittest.TestCase):
    def setUp(self):
        self.mipila = pila()
        self.mipila.add_item(10);
        self.mipila.add_item(20);
        self.mipila.add_item(22, 33);

    def test_flow(self):
        self.assertEqual(self.mipila.pop_item(), 33)
        self.assertEqual(self.mipila.pop_item(), 22)
        self.assertEqual(self.mipila.count_items(), 2)
        while self.mipila.pop_item(): pass

        self.assertEqual(self.mipila.count_items(), 0)

if __name__ == "__main__" :
    unittest.main()